Refactor provider integration tests
