+++ bxsolana/provider/constants.py
@@ -23,3 +23,5 @@ LOCAL_API_HTTP = "http://127.0.0.1:9000"
 LOCAL_API_WS = "ws://127.0.0.1:9000/ws"
 LOCAL_API_GRPC_HOST = "127.0.0.1"
 LOCAL_API_GRPC_PORT = 9000
+
+
