+++ bxsolana/utils/constants.py
@@ -1,0 +1,8 @@
+# sample keys to run integration/regression tests with
+# maintained by bloxRoute team
+PUBLIC_KEY = "4JCZomAb4eKcJQ9fqoi2JPT1TzjbhJEk8V8MmFKUQfnV"
+USDC_WALLET = "8w1igsiZfsWnMTJSKkSvDQYYCZyf6aXMihuXEYpxcZVD"
+OPEN_ORDERS = "4JCZomAb4eKcJQ9fqoi2JPT1TzjbhJEk8V8MmFKUQfnV"
+SOL_USDC_MARKET = "9wFFyRfZBsuAha4YcuxcXLKwMxJR43S7fPfQLusDBzvT"
+ORDER_ID = "3929156487700134707538850"
+
